requirements.txt
sounddevice
moviepy
mplayer.py
